requirements.txt
streamlit
sentence-transformers
faiss-cpu
requests
numpy